ai-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
spacy>=3.7.2
openai>=1.10.0
PyMuPDF>=1.23.8
python-docx>=1.1.0
python-dotenv>=1.0.1
